requirements.txt
requests==2.31.0
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.0.1
streamlit
pandas